requirements.txt
Django==1.10.7
django-extensions==1.7.9
django-filer==1.2.7
django-mptt==0.8.7
django-polymorphic==1.0.2
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.6.3
easy-thumbnails==2.3
olefile==0.44
Pillow==4.1.1
pytz==2017.2
six==1.10.0
unicode-slugify==0.1.3
Unidecode==0.4.20
